Extract HttpRSocketFunction to a separate class.
